+++ tests/Unit/Document/ResourceObjectTest.php
@@ -81,6 +81,9 @@ class ResourceObjectTest extends TestCase
                         'related' => 'http://localhost/api/v1/posts/1/tags',
                         'self' => 'http://localhost/api/v1/posts/1/relationships/tags',
                     ],
+                    'meta' => [
+                        'count' => 2,
+                    ],
                 ],
             ],
             'links' => [
@@ -620,6 +623,50 @@ class ResourceObjectTest extends TestCase
         $fields = $resource->all();
 
         $this->assertArrayNotHasKey('comments', $fields);
+    }
+
+    public function testWithoutMeta(): void
+    {
+        $this->values['meta'] = ['foo' => 'bar'];
+        $this->resource = ResourceObject::fromArray($this->values);
+
+        $expected = $this->values;
+
+        unset(
+            $expected['meta'],
+            $expected['relationships']['tags']['meta'],
+        );
+
+        $this->assertNotSame($this->resource, $resource = $this->resource->withoutMeta());
+
+        $this->assertSame($this->values, $this->resource->jsonSerialize());
+        $this->assertSame($expected, $resource->jsonSerialize());
+    }
+
+    public function testWithRelationshipMeta(): void
+    {
+        $expected = $this->values;
+        $expected['relationships']['comments']['meta'] = ['count' => 5];
+
+        $actual = $this->resource->withRelationshipMeta('comments', ['count' => 5]);
+
+        $this->assertNotSame($this->resource, $actual);
+        $this->assertSame($this->values, $this->resource->jsonSerialize());
+        $this->assertSame($expected, $actual->jsonSerialize());
+    }
+
+    public function testWithRelationshipMetaOnNewRelationship(): void
+    {
+        $expected = $this->values;
+        $expected['relationships']['likes'] = [];
+        $expected['relationships']['likes']['meta'] = ['count' => 5];
+        ksort($expected['relationships']);
+
+        $actual = $this->resource->withRelationshipMeta('likes', ['count' => 5]);
+
+        $this->assertNotSame($this->resource, $actual);
+        $this->assertSame($this->values, $this->resource->jsonSerialize());
+        $this->assertSame($expected, $actual->jsonSerialize());
     }
 
     public function testMerge(): void